server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
Pillow==10.3.0
aiofiles==23.2.1
httpx==0.27.0
websockets==12.0
slowapi==0.1.9
emails==0.6
jinja2==3.1.4
redis==5.0.4
celery==5.4.0
